+++ src/index.ts
@@ -1,7 +1,7 @@
+export { default as crawlFavicons } from './crawlFavicons';
+export { default as crawlWithFetch } from './crawlWithFetch';
+export { default as crawlWithPuppeteer } from './crawlWithPuppeteer';
 export { default as findFavicons } from './findFavicons';
-export { default as fetchFavicons } from './fetchFavicons';
-export { default as fetchStringWithNodeFetch } from './fetchStringWithNodeFetch';
-export { default as fetchStringWithPuppeteer } from './fetchStringWithPuppeteer';
 export * from './constants';
 export * from './lib';
 export * from './types';
